libfetch: allow to optionally verify CRL with SSL_CRL_OPTIONAL
This works with both SSL_CRL_FILE and SSL_CRL_VERIFY.

The situation is not ideal, but since we do not know what we are
going to deal with in terms of random Internet hosts an optional
knob is needed, e.g. pkg multi-repo cases on the user side.  For
the benefit of the doubt print the appropriate message for the
user to see.